+++ components/archive/impl/src/main/java/com/flipperdevices/archive/impl/viewmodel/GeneralTabViewModel.kt
@@ -1,6 +1,5 @@
 package com.flipperdevices.archive.impl.viewmodel
 
-import androidx.lifecycle.viewModelScope
 import com.flipperdevices.bridge.dao.api.delegates.FavoriteApi
 import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi
 import com.flipperdevices.bridge.dao.api.model.FlipperKey