+++ components/infrared/editor/src/main/kotlin/com/flipperdevices/infrared/editor/viewmodel/InfraredEditorViewModel.kt
@@ -3,7 +3,6 @@ package com.flipperdevices.infrared.editor.viewmodel
 import android.content.Context
 import android.os.Vibrator
 import androidx.core.content.ContextCompat
-import androidx.lifecycle.viewModelScope
 import com.flipperdevices.bridge.api.utils.FlipperSymbolFilter
 import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi
 import com.flipperdevices.bridge.dao.api.delegates.key.UpdateKeyApi