+++ components/widget/screen/src/main/java/com/flipperdevices/widget/screen/viewmodel/WidgetSelectViewModel.kt
@@ -4,7 +4,6 @@ import android.app.Activity
 import android.app.Activity.RESULT_OK
 import android.appwidget.AppWidgetManager
 import android.content.Intent
-import androidx.lifecycle.viewModelScope
 import com.flipperdevices.bridge.dao.api.delegates.FavoriteApi
 import com.flipperdevices.bridge.dao.api.delegates.WidgetDataApi
 import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi